refactor : error 패키지 제거
