+++ src/main/java/com/yapp/artie/domain/user/exception/InvalidUidException.java
@@ -1,11 +1,11 @@
 package com.yapp.artie.domain.user.exception;
 
-import com.yapp.artie.global.error.exception.common.BusinessException;
-import com.yapp.artie.global.error.response.ErrorCode;
+import com.yapp.artie.global.exception.common.BusinessException;
+import com.yapp.artie.global.exception.response.ErrorCode;
 
 public class InvalidUidException extends BusinessException {
 
   public InvalidUidException() {
-    super(ErrorCode.AUTH_INVALID_USERINFO);
+    super(ErrorCode.FIREBASE_INVALID_USERINFO);
   }
 }
