+++ src/main/java/com/yapp/artie/global/authentication/JwtService.java
@@ -3,8 +3,8 @@ package com.yapp.artie.global.authentication;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthException;
 import com.google.firebase.auth.FirebaseToken;
-import com.yapp.artie.global.error.exception.authentication.InvalidTokenException;
-import com.yapp.artie.global.error.exception.authentication.NotExistValidTokenException;
+import com.yapp.artie.global.exception.authentication.InvalidTokenException;
+import com.yapp.artie.global.exception.authentication.NotExistValidTokenException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
