+++ src/main/java/com/yapp/artie/global/exception/GlobalExceptionHandler.java
@@ -1,8 +1,8 @@
-package com.yapp.artie.global.error.handler;
+package com.yapp.artie.global.exception;
 
-import com.yapp.artie.global.error.response.ErrorCode;
-import com.yapp.artie.global.error.response.ErrorResponse;
-import com.yapp.artie.global.error.exception.common.BusinessException;
+import com.yapp.artie.global.exception.response.ErrorCode;
+import com.yapp.artie.global.exception.response.ErrorResponse;
+import com.yapp.artie.global.exception.common.BusinessException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
